Add support for stories having a year associated with them
When a story is created or edited a four digit year may be added. This year
will be displayed in parens in a small font after the story title. If a
story is submitted with " (YYYY)" at the end of the title it will be turned
into a stored year (and the title edited) on reload.